cleanup mkdocs for pandas general functions
