+++ janitor/functions/__init__.py
@@ -2,18 +2,18 @@
 # General Functions
 
 pyjanitor's general-purpose data cleaning functions.
-
-NOTE: Instructions for future contributors:
-
-1. Place the source code of the functions in a file named after the function.
-2. Place utility functions in the same file.
-3. If you use a utility function from another source file,
-please refactor it out to `janitor.functions.utils`.
-4. Import the function into this file so that it shows up in the top-level API.
-5. Sort the imports in alphabetical order.
-6. Try to group related functions together (e.g. see `convert_date.py`)
-7. Never import utils.
 """
+
+# NOTE: Instructions for future contributors:
+
+# 1. Place the source code of the functions in a file named after the function.
+# 2. Place utility functions in the same file.
+# 3. If you use a utility function from another source file,
+# please refactor it out to `janitor.functions.utils`.
+# 4. Import the function into this file so that it shows up in the top-level API.
+# 5. Sort the imports in alphabetical order.
+# 6. Try to group related functions together (e.g. see `convert_date.py`)
+# 7. Never import utils.
 
 from .add_columns import add_columns
 from .also import also
@@ -65,7 +65,13 @@ from .rename_columns import rename_column, rename_columns
 from .reorder_columns import reorder_columns
 from .round_to_fraction import round_to_fraction
 from .row_to_names import row_to_names
-from .select import select, select_columns, select_rows
+from .select import (
+    get_columns,
+    get_index_labels,
+    select,
+    select_columns,
+    select_rows,
+)
 from .shuffle import shuffle
 from .sort_column_value_order import sort_column_value_order
 from .sort_naturally import sort_naturally
@@ -78,10 +84,6 @@ from .truncate_datetime import truncate_datetime_dataframe
 from .update_where import update_where
 from .utils import (
     DropLabel,
-    col,
-    get_columns,
-    get_index_labels,
-    patterns,
     unionize_dataframe_categories,
 )
 
@@ -156,10 +158,8 @@ __all__ = [
     "transform_columns",
     "truncate_datetime_dataframe",
     "update_where",
-    "patterns",
     "unionize_dataframe_categories",
     "DropLabel",
     "get_index_labels",
-    "col",
     "get_columns",
 ]
